+++ src/main/java/tube/jdbc/UserJDBCTemplate.java
@@ -53,7 +53,7 @@ public class UserJDBCTemplate{
 			 public User mapRow(ResultSet rs, int rowNum) throws SQLException {
 
 				User user = new User();
-				user.setId(rs.getInt("id"));
+				user.setId(rs.getLong("id"));
 				user.setUsername(rs.getString("username"));
 				user.setPassword(rs.getString("password"));
 				user.setEmail(rs.getString("email"));
