+++ src/main/java/tube/jdbc/UserMapper.java
@@ -10,7 +10,7 @@ import tube.entities.User;
 public class UserMapper implements RowMapper<User> {
 	public User mapRow(ResultSet rs, int rowNum) throws SQLException {
 		User user = null;
-		user = new User(rs.getInt("id"), rs.getString("username"), rs.getString("email"), rs.getString("password"));
+		user = new User(rs.getLong("id"), rs.getString("username"), rs.getString("email"), rs.getString("password"));
 		return user;
 	}
 }
