+++ src/main/java/tube/jdbc/VideoMapper.java
@@ -11,7 +11,7 @@ public class VideoMapper implements RowMapper<Video> {
 
 	public Video mapRow(ResultSet rs, int rowNum) throws SQLException {
 		Video video = null;
-		video = new Video(rs.getInt("id"), rs.getString("description"), rs.getString("file_name"), rs.getString("title"), rs.getInt("user_id"));
+		video = new Video(rs.getInt("id"), rs.getString("description"), rs.getString("file_name"), rs.getString("title"), rs.getLong("user_id"));
 		return video;
 	}
 }
